requirements.txt
Flask==3.1.3
google-generativeai==0.8.6
numpy==2.4.0
pandas==3.0.1
pypdf==6.4.2
python-dotenv==1.2.2
scikit-learn==1.8.0
pytest==8.4.2
